+++ changelog.md
@@ -12,6 +12,9 @@
 * Add `powerSet`, `cartesianProduct`, and `disjointUnion` for
   `Data.Set` (Thanks, Edward Kmett!)
 
+* Make `Data.Sequence.replicateM` a synonym for `replicateA`
+  for post-AMP `base`.
+
 * Make `>>=` for `Data.Tree` strict in the result of its second argument;
   being too lazy here is almost useless, and violates one of the monad identity
   laws. Specifically, `return () >>= \_ -> undefined` should always be